1) Fixed swapping kernel arguments; 2) Transfer to device empty buffer each 32 cycles
1) Corrected the swapping of the kernel arguments when total_elements to compute
is bigger then zero. The mistake was that arguments were swapped by following the
external loop index, introducing so the wrong swap sequence. Introduced a new
variable that takes in account this issue.

2) Each 32 cycle after transferring the output buffer from the device to the host,
right after a new empty buffer should be sent back to device in order to avoid
storing the new results to the old buffer.